io: split enigma, jam, and kin into their own modules
