+++ lorenz-io-enigma/gradle.properties
@@ -1,0 +1,4 @@
+name = Lorenz-IO-Enigma
+description = An implementation of the Enigma mapping format for Lorenz.
+url = https://www.jamiemansfield.me/projects/lorenz
+inceptionYear = 2018
