+++ lorenz-io-enigma/src/main/java/org/cadixdev/lorenz/io/enigma/EnigmaConstants.java
