+++ lorenz-io-enigma/src/test/resources/test.enigma
