+++ lorenz-io-jam/gradle.properties
@@ -1,0 +1,4 @@
+name = Lorenz-IO-JAM
+description = An implementation of the JAM mapping format for Lorenz.
+url = https://www.jamiemansfield.me/projects/lorenz
+inceptionYear = 2018
