+++ lorenz-io-jam/src/main/java/org/cadixdev/lorenz/io/jam/JamConstants.java
