+++ lorenz-io-jam/src/main/java/org/cadixdev/lorenz/io/jam/JamMappingFormat.java
