+++ lorenz-io-kin/gradle.properties
@@ -1,0 +1,4 @@
+name = Lorenz-IO-Kin
+description = An implementation of the Kin mapping format for Lorenz.
+url = https://www.jamiemansfield.me/projects/lorenz
+inceptionYear = 2018
