+++ lorenz-io-kin/src/main/java/org/cadixdev/lorenz/io/kin/KinConstants.java
