+++ lorenz-io-kin/src/main/java/org/cadixdev/lorenz/io/kin/KinMappingFormat.java
