+++ lorenz-io-kin/src/main/java/org/cadixdev/lorenz/io/kin/KinReader.java
@@ -29,13 +29,13 @@ import static org.cadixdev.lorenz.io.kin.KinConstants.MAGIC;
 import static org.cadixdev.lorenz.io.kin.KinConstants.VERSION_ONE;
 import static org.cadixdev.lorenz.io.kin.KinConstants.toHexString;
 
+import org.cadixdev.bombe.type.FieldType;
+import org.cadixdev.bombe.type.signature.FieldSignature;
 import org.cadixdev.lorenz.MappingSet;
 import org.cadixdev.lorenz.io.BinaryMappingsReader;
 import org.cadixdev.lorenz.io.MappingsReader;
 import org.cadixdev.lorenz.model.ClassMapping;
 import org.cadixdev.lorenz.model.FieldMapping;
-import org.cadixdev.bombe.type.FieldType;
-import org.cadixdev.bombe.type.signature.FieldSignature;
 
 import java.io.IOException;
 import java.io.InputStream;
